Fixed FoxmlDocument not respecting versioning
FoxmlDocument wasn't correcly setting the versioning
flag on datastreams. This was causing any object created
in Tuque using FoxmlDocument to have its datastreams set
to non-versionable.

The fix was a simple error serializing the output of a
boolean to XML. PHP was outputting 1 or 0 for true or
false, while the foxml was expecting true or false. Any
other value was interperated as false.